requirements.txt
beautifulsoup4==4.8.1
config==0.4.2
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11.23
django-bootstrap3==11.1.0
django-bootstrap4==1.0.1
django-heroku==0.3.1
django-phone-field==1.8.0
django-registration==2.4.1
django-tinymce==2.8.0
djangorestframework==3.10.3
gunicorn==19.9.0
Pillow==6.2.0
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.3
soupsieve==1.9.4
whitenoise==4.1.4
